+++ packages/daobox-use-aragon/src/lib/index.ts
@@ -1,3 +1,6 @@
 export * from "./validatData";
 export * from "./regex";
 export * from "./validators";
+export * from "./hexTools";
+export * from "./setQueryKey";
+export * from "./convertChain";